review: address pr feedback
