+++ packages/imperative/src/imperative/src/plugins/cmd/install/install.definition.ts
@@ -12,70 +12,7 @@
 import { ICommandDefinition } from "../../../../../cmd";
 import { join } from "path";
 import { PMFConstants } from "../../utilities/PMFConstants";
-import { ImperativeConfig } from "../../../../../utilities";
-
-const cliCmdName = ImperativeConfig.instance.findPackageBinName() ?
-    ImperativeConfig.instance.findPackageBinName() : "Your_CLI_Command_Name";
-
-const pluginDescription =
-    "A space-separated list of plug-ins to install. A plug-in can be " +
-    "any format that is accepted by the `npm install` command (local directory, " +
-    "TAR file, git URL, public package, private package, etc...).\n" +
-    "\n" +
-    "To use a relative local directory, at least one '/' or '\\' must exist in " +
-    "the plug-in path. For example, you have a local plug-in in a folder called " +
-    "'test-plugin' that you want to install. Specify the relative local directory " +
-    "by issuing the following command:\n\n" +
-    `${cliCmdName} plugins install ./test-plugin\n\n` +
-    "If you omit the './', then the install command looks for 'test-plugin' in an " +
-    "npm registry.\n" +
-    "\n" +
-    "If the plugin argument is omitted, the plugins.json file will determine which " +
-    "plug-ins are installed. For more information on the plugins.json file, see the --file option.";
-
-const registryDescription =
-    "The npm registry that is used when installing remote packages. When this value is omitted, the " +
-    "value returned by `npm config get registry` is used.\n" +
-    "\n" +
-    "For more information about npm registries, see: " +
-    "https://docs.npmjs.com/misc/registry";
-
-const fileDescription =
-    "Specifies the location of a plugins.json file that contains the plug-ins you want to install.\n" +
-    "\n" +
-    "All plug-ins specified in plugins.json will be installed to the base CLI and " +
-    `the contents will be placed into ${PMFConstants.instance.PLUGIN_JSON}.\n` +
-    "\n" +
-    "If you do not specify a plugins.json file and do not specify a plug-in, the default " +
-    `plugin.json file (${PMFConstants.instance.PLUGIN_JSON}) will be used. This provides a ` +
-    "way to install plug-ins that were lost or corrupted after " +
-    `reinstalling or updating ${ImperativeConfig.instance.loadedConfig.productDisplayName}.`;
-
-const loginDescription =
-    "The flag to add a registry user account to install from secure registry. It saves credentials " +
-    "to the .npmrc file using `npm login`. When this value is omitted, credentials from .npmrc file is used. " +
-    "If you used this flag once for specific registry, you don't have to use it again, it uses credentials from .npmrc file.\n" +
-    "\n" +
-    "For more information about npm registries, see: \n" +
-    "https://docs.npmjs.com/cli/login for NPM >= 9\n" +
-    "https://docs.npmjs.com/cli/adduser for NPM < 9";
-
-const verboseDescription =
-    "Specifies that verbose output is printed for npm install.\n" +
-    "This may be useful for debugging errors during plugin installation.";
-
-const allowScriptsDescription =
-    "A comma-separated list of package names that are allowed to run their npm install scripts " +
-    "(preinstall, install, postinstall, and prepare). The names must match exactly. Zowe passes " +
-    "this list to npm as its own --allow-scripts option.\n" +
-    "\n" +
-    "npm 12 blocks these scripts unless the package is in the list. This breaks plug-ins that " +
-    "have a dependency that must be built during install, such as the Db2 plug-in and its " +
-    "ibm_db dependency. Older versions of npm do not know this option and ignore it, so using " +
-    "this option does not change how they work. If you do not use this option, Zowe does not " +
-    "pass anything to npm, and npm decides what to do.\n" +
-    "\n" +
-    "Only list packages that you trust. Install scripts can run any code on your machine.";
+import { PluginCmdConstants } from "../PluginCmdConstants";
 
 /**
  * Definition of the install command.
@@ -91,7 +28,7 @@ export const installDefinition: ICommandDefinition = {
         {
             name: "plugin...",
             type: "string",
-            description: pluginDescription,
+            description: PluginCmdConstants.INSTALL_PLUGIN_DESCRIPTION,
             required: false
         }
     ],
@@ -99,33 +36,33 @@ export const installDefinition: ICommandDefinition = {
         {
             name: "file",
             type: "existingLocalFile",
-            description: fileDescription,
+            description: PluginCmdConstants.FILE_DESCRIPTION,
             required: false,
             conflictsWith: ["registry"]
         },
         {
             name: "registry",
             type: "string",
-            description: registryDescription,
+            description: PluginCmdConstants.REGISTRY_DESCRIPTION,
             required: false
         },
         {
             name: "login",
             type: "boolean",
-            description: loginDescription,
+            description: PluginCmdConstants.LOGIN_DESCRIPTION,
             required: false,
             implies: ["registry"]
         },
         {
             name: "verbose",
             type: "boolean",
-            description: verboseDescription,
+            description: PluginCmdConstants.VERBOSE_DESCRIPTION,
             required: false
         },
         {
             name: "allow-scripts",
             type: "string",
-            description: allowScriptsDescription,
+            description: PluginCmdConstants.ALLOW_SCRIPTS_DESCRIPTION,
             required: false
         }
     ],
